requirements.txt
numpy
pandas
python-dotenv
#-e .
mysql-connector-python
pymysql
scikit-learn
cryptography